LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/parsing/overflow-clip-margin-computed-expected.txt


PASS Property overflow-clip-margin value '0px'
FAIL Property overflow-clip-margin value '10px' assert_equals: expected "10px" but got "0px"
FAIL Property overflow-clip-margin value 'content-box' assert_equals: expected "content-box" but got "0px"
FAIL Property overflow-clip-margin value 'content-box 0px' assert_equals: expected "content-box" but got "0px"
FAIL Property overflow-clip-margin value 'content-box 10px' assert_equals: expected "content-box 10px" but got "0px"
FAIL Property overflow-clip-margin value '10px content-box' assert_equals: expected "content-box 10px" but got "0px"
PASS Property overflow-clip-margin value 'padding-box'
PASS Property overflow-clip-margin value 'padding-box 0px'
FAIL Property overflow-clip-margin value 'padding-box 10px' assert_equals: expected "10px" but got "0px"
FAIL Property overflow-clip-margin value '10px padding-box' assert_equals: expected "10px" but got "0px"
FAIL Property overflow-clip-margin value 'border-box' assert_equals: expected "border-box" but got "0px"
FAIL Property overflow-clip-margin value 'border-box 0px' assert_equals: expected "border-box" but got "0px"
FAIL Property overflow-clip-margin value 'border-box 10px' assert_equals: expected "border-box 10px" but got "0px"
FAIL Property overflow-clip-margin value '10px border-box' assert_equals: expected "border-box 10px" but got "0px"
